+++ extensions/Worker.Extensions.CosmosDB/release_notes.md
@@ -4,7 +4,8 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.8.0
+### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.8.1
 
-- Updating `Microsoft.Azure.WebJobs.Extensions.CosmosDB` reference to 4.6.0
-- Updating `Microsoft.Extensions.Azure` reference to 1.7.2
+- Updating `Microsoft.Azure.WebJobs.Extensions.CosmosDB` reference to 4.6.1
+- Updating `Microsoft.Extensions.Azure` reference to 1.7.3
+- Updating `Microsoft.Azure.Cosmos` reference to 3.39.1